move: 결과출력 기능 InputView to Outputview
printResult, printWinners, printCurrentCarLocation 메서드 이동
